Move evmc.h to include/evmc dir
