+++ examples/examplevm/examplevm.h
@@ -5,7 +5,7 @@
 
 #pragma once
 
-#include <evmc.h>
+#include <evmc/evmc.h>
 
 /**
  * Creates EVMC Example VM.
