+++ test/vmtester/vmtester.cpp
@@ -1,12 +1,8 @@
-// EVMC -- Ethereum Client-VM Connector API
+// EVMC: Ethereum Client-VM Connector API.
 // Copyright 2018 Pawel Bylica.
 // Licensed under the MIT License. See the LICENSE file.
 
 #include "vmtester.hpp"
-
-#include <evmc.h>
-
-#include <gtest/gtest.h>
 
 #include <boost/dll.hpp>
 #include <boost/dll/alias.hpp>
